requirements.txt
django
djangorestframework
python-slugify
psycopg2-binary
Pillow
drf-yasg
django-cors-headers
python-decouple
django-filter

djangorestframework-simplejwt
celery
redis
